statistics

+++ CoreLogic/Services/Blockchain/Ethereum/IEthereumReader.cs
@@ -9,7 +9,7 @@ namespace Etherama.CoreLogic.Services.Blockchain.Ethereum {
 		/// <summary>
 		/// Get latest block number on the logs provider side
 		/// </summary>
-		Task<BigInteger> GetLogsLatestBlockNumber();
+		Task<long> GetLogsLatestBlockNumber();
 
 		/// <summary>
 		/// Check chain transaction by it's ID
@@ -17,7 +17,7 @@ namespace Etherama.CoreLogic.Services.Blockchain.Ethereum {
 		Task<TransactionInfo> CheckTransaction(string txid, int confirmationsRequired);
 
 		/// <summary>
-		/// Get current gas price
+		/// Get current gas price in wei
 		/// </summary>
 		Task<BigInteger> GetCurrentGasPrice();
